RDKBEMU-247 : Unknown Code found in rdkbemu_xb3
Reason for change: Deleting the Unknown Code found in
"webui/source/Styles/xb3/code/includes/utility.php" as it is not used.
Test Procedure: Test from GUI,Unknown Code found is in comments and it is removed.
Please test and see if the pages mentioned below are working fine.
	connected_devices_computers.php
	managed_sites.php
	port_forwarding.php
	port_triggering.php
	parental_reports.php
	troubleshooting_logs.php
	dmz.php
Risks: None
Signed-off-by: rnarayanan <rajkumar.narayanan@lnttechservices.com>

Change-Id: I1ee0eec1e34549506771e0f4c1d4d212b0b2c5ad